Update OpenSchedule.java
Alterações na classe OpenSchedule onde agora possui botões que permitem ver as semanas.
JText temporário
