+++ horarioSemana.json
@@ -1,181 +1,1 @@
-[ {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP04",
-  "Turma" : "ET-A9, ET-A8, ET-A7, ET-A12, ET-A11, ET-A10",
-  "Inscritos no turno" : 48,
-  "Dia da semana" : "Ter",
-  "Hora início da aula" : "13:00:00",
-  "Hora fim da aula" : "14:30:00",
-  "Data da aula" : "13/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP01",
-  "Turma" : "ET-A6, ET-A5, ET-A4, ET-A3, ET-A2, ET-A1",
-  "Inscritos no turno" : 37,
-  "Dia da semana" : "Ter",
-  "Hora início da aula" : "14:30:00",
-  "Hora fim da aula" : "16:00:00",
-  "Data da aula" : "13/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP05",
-  "Turma" : "EI-PL-A2, EI-PL-A1",
-  "Inscritos no turno" : 83,
-  "Dia da semana" : "Ter",
-  "Hora início da aula" : "19:30:00",
-  "Hora fim da aula" : "21:00:00",
-  "Data da aula" : "13/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP05",
-  "Turma" : "EI-PL-A2, EI-PL-A1",
-  "Inscritos no turno" : 83,
-  "Dia da semana" : "Ter",
-  "Hora início da aula" : "21:00:00",
-  "Hora fim da aula" : "22:30:00",
-  "Data da aula" : "13/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP02",
-  "Turma" : "EI-A6, EI-A5, EI-A4, EI-A3, EI-A2, EI-A1",
-  "Inscritos no turno" : 52,
-  "Dia da semana" : "Qua",
-  "Hora início da aula" : "13:00:00",
-  "Hora fim da aula" : "14:30:00",
-  "Data da aula" : "14/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP03",
-  "Turma" : "EI-A9, EI-A8, EI-A7, EI-A12, EI-A11, EI-A10",
-  "Inscritos no turno" : 53,
-  "Dia da semana" : "Qua",
-  "Hora início da aula" : "14:30:00",
-  "Hora fim da aula" : "16:00:00",
-  "Data da aula" : "14/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP01",
-  "Turma" : "ET-A6, ET-A5, ET-A4, ET-A3, ET-A2, ET-A1",
-  "Inscritos no turno" : 37,
-  "Dia da semana" : "Qui",
-  "Hora início da aula" : "13:00:00",
-  "Hora fim da aula" : "14:30:00",
-  "Data da aula" : "15/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP03",
-  "Turma" : "EI-A9, EI-A8, EI-A7, EI-A12, EI-A11, EI-A10",
-  "Inscritos no turno" : 53,
-  "Dia da semana" : "Qui",
-  "Hora início da aula" : "13:00:00",
-  "Hora fim da aula" : "14:30:00",
-  "Data da aula" : "15/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP02",
-  "Turma" : "EI-A6, EI-A5, EI-A4, EI-A3, EI-A2, EI-A1",
-  "Inscritos no turno" : 52,
-  "Dia da semana" : "Qui",
-  "Hora início da aula" : "14:30:00",
-  "Hora fim da aula" : "16:00:00",
-  "Data da aula" : "15/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP04",
-  "Turma" : "ET-A9, ET-A8, ET-A7, ET-A12, ET-A11, ET-A10",
-  "Inscritos no turno" : 48,
-  "Dia da semana" : "Qui",
-  "Hora início da aula" : "14:30:00",
-  "Hora fim da aula" : "16:00:00",
-  "Data da aula" : "15/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP05",
-  "Turma" : "EI-PL-A2, EI-PL-A1",
-  "Inscritos no turno" : 83,
-  "Dia da semana" : "Qui",
-  "Hora início da aula" : "19:30:00",
-  "Hora fim da aula" : "21:00:00",
-  "Data da aula" : "15/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP03",
-  "Turma" : "EI-A9, EI-A8, EI-A7, EI-A12, EI-A11, EI-A10",
-  "Inscritos no turno" : 53,
-  "Dia da semana" : "Sex",
-  "Hora início da aula" : "13:00:00",
-  "Hora fim da aula" : "14:30:00",
-  "Data da aula" : "16/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP02",
-  "Turma" : "EI-A6, EI-A5, EI-A4, EI-A3, EI-A2, EI-A1",
-  "Inscritos no turno" : 52,
-  "Dia da semana" : "Sex",
-  "Hora início da aula" : "14:30:00",
-  "Hora fim da aula" : "16:00:00",
-  "Data da aula" : "16/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP01",
-  "Turma" : "ET-A6, ET-A5, ET-A4, ET-A3, ET-A2, ET-A1",
-  "Inscritos no turno" : 37,
-  "Dia da semana" : "Sex",
-  "Hora início da aula" : "14:30:00",
-  "Hora fim da aula" : "16:00:00",
-  "Data da aula" : "16/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-}, {
-  "Curso" : "LETI, LEI, LEI-PL",
-  "Unidade Curricular" : "Cálculo I",
-  "Turno" : "03703TP04",
-  "Turma" : "ET-A9, ET-A8, ET-A7, ET-A12, ET-A11, ET-A10",
-  "Inscritos no turno" : 48,
-  "Dia da semana" : "Sex",
-  "Hora início da aula" : "16:00:00",
-  "Hora fim da aula" : "17:30:00",
-  "Data da aula" : "16/09/2022",
-  "Sala atribuída à aula" : "",
-  "Lotação da sala" : 0
-} ]
+[ ]
