+++ src/main/java/softwareeng/project/OpenSchedule.java
@@ -1,23 +1,37 @@
 package softwareeng.project;
 
 
+import com.opencsv.exceptions.CsvValidationException;
+
 import javax.swing.*;
 import javax.swing.filechooser.FileNameExtensionFilter;
 import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
+import java.util.List;
 import java.util.logging.Logger;
 
 
 public class OpenSchedule extends JFrame {
     private JButton csvButton;
     private JButton jsonButton;
+
+    private JButton selectFileButton;
+
+    private Horario h;
+
+    private String filename;
     private JButton backButton;
 
+    private JPanel panel;
+
     private static final Logger LOGGER = Logger.getLogger("OpenSchedule");
+
+    private static final String ERROR_MESSAGE = "Error";
 
 
     public OpenSchedule(){
@@ -51,38 +65,26 @@ public class OpenSchedule extends JFrame {
     }
 
     private void  initComponents(){
-        csvButton = new JButton("Open CSV File");
-        jsonButton = new JButton("Open Json File");
-        csvButton.setBorderPainted(false);
-        jsonButton.setBorderPainted(false);
-        csvButton.setFocusPainted(false);
-        jsonButton.setFocusPainted(false);
-        csvButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-        jsonButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+        int buttonSize = 48;
+        selectFileButton = new JButton("Convert your File");
+        selectFileButton.setBorderPainted(false);
+        selectFileButton.setFocusPainted(false);
+        selectFileButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+        selectFileButton.setIcon(new ImageIcon(new ImageIcon("icons/convertHorario.png")
+                .getImage().getScaledInstance(buttonSize, buttonSize, java.awt.Image.SCALE_SMOOTH)));
         backButton = new JButton();        backButton.setBorderPainted(false);
         backButton.setFocusPainted(false);
         backButton.setBackground(new Color(0, 0, 0, 0)); // Set transparent background
         backButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-
-        int buttonSize = 48;
-        csvButton.setIcon(new ImageIcon(new ImageIcon("icons/CSV.png")
-                .getImage().getScaledInstance(buttonSize, buttonSize, java.awt.Image.SCALE_SMOOTH)));
-        jsonButton.setIcon(new ImageIcon(new ImageIcon("icons/Json.png")
-                .getImage().getScaledInstance(buttonSize, buttonSize, java.awt.Image.SCALE_SMOOTH)));
         backButton.setIcon(new ImageIcon(new ImageIcon("icons/back.png")
                 .getImage().getScaledInstance(buttonSize, buttonSize, java.awt.Image.SCALE_SMOOTH)));
     }
 
     private void layoutComponents() {
         JPanel rowOnePanel = new JPanel(new GridLayout(1, 2));
-        rowOnePanel.add(csvButton);
-        rowOnePanel.add(csvButton);
+        rowOnePanel.add(selectFileButton);
         add(rowOnePanel);
 
-        JPanel rowTwoPanel = new JPanel(new GridLayout(1, 2));
-        rowTwoPanel.add(jsonButton);
-        rowTwoPanel.add(jsonButton);
-        add(rowTwoPanel);
 
         JPanel rowThreePanel = new JPanel(new GridBagLayout());
         GridBagConstraints gbc = new GridBagConstraints();
@@ -96,94 +98,80 @@ public class OpenSchedule extends JFrame {
     }
 
     private void addListeners() {
-        csvButton.addActionListener(e -> openCSVSchedule());
-        jsonButton.addActionListener(e -> openJSonSchedule());
+
+        selectFileButton.addActionListener(e -> {
+            try {
+                selectFileButtonClicked();
+            } catch (CsvValidationException | IOException ex) {
+                JOptionPane.showMessageDialog(this, ex.getMessage(), ERROR_MESSAGE, JOptionPane.ERROR_MESSAGE);
+            }
+        });
         backButton.addActionListener(e -> backToMainMenu());
     }
 
-    private void openCSVSchedule(){
+
+    private void selectFileButtonClicked() throws CsvValidationException, IOException {
         JFileChooser fileChooser = new JFileChooser();
         fileChooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
-        fileChooser.setFileFilter(new FileNameExtensionFilter("CSV files", "csv"));
+
+        FileNameExtensionFilter csvFilter = new FileNameExtensionFilter("CSV files (*.csv)", "csv");
+        fileChooser.setFileFilter(csvFilter);
+
+        FileNameExtensionFilter jsonFilter = new FileNameExtensionFilter("JSON files (*.json)", "json");
+        fileChooser.setFileFilter(jsonFilter);
+
 
         int result = fileChooser.showOpenDialog(this);
         if (result == JFileChooser.APPROVE_OPTION) {
+            String path = fileChooser.getSelectedFile().getPath();
             File selectedFile = fileChooser.getSelectedFile();
+            if (selectedFile.getName().endsWith(".csv") || selectedFile.getName().endsWith(".json")) {
+                selectFileButton.setVisible(false);
+                filename = selectedFile.getName();
+                h = new Horario(filename);
+                h.getWeek(selectedFile.getName(), 1);
+                //Exibir horário
 
-            if (selectedFile.getName().endsWith(".csv")) {
-                openCSV(selectedFile);
-            } else {
-                JOptionPane.showMessageDialog(this, "The selected file is not a CSV file.", "Error", JOptionPane.ERROR_MESSAGE);
-                openCSVSchedule(); // restart the file chooser
+                List<Session> session = h.converFileToArray("horarioSemana.json");
+
+
+
+                panel = new JPanel();
+                JTextField textField = new JTextField(session.get(0).toString()); // Temporário
+                panel.add(textField);
+                add(panel);
+
+
+                int numberWeeks = h.countWeeks(filename);
+                JPanel buttonPanel = new JPanel();
+                for (int count = 1; count <= numberWeeks; count++) {
+                    JButton button = new JButton(Integer.toString(count));
+                    button.addActionListener(new ActionListener() {
+                        public void actionPerformed(ActionEvent e) {
+                            String buttonText = button.getText();
+                            h.getWeek(filename, Integer.parseInt(buttonText));
+                            List<Session> session = h.converFileToArray("horarioSemana.json");
+                            if (session.isEmpty()) {
+                                JOptionPane.showMessageDialog(buttonPanel, "Não possui aulas marcadas nessa semana");
+
+
+                            } else {
+
+                                textField.setText(session.get(0).toString());
+
+                            }
+                        }
+                    });
+                    buttonPanel.add(button);
+                }
+                add(buttonPanel, BorderLayout.SOUTH);
             }
+
         }
     }
 
-    private void openJSonSchedule(){
-        JFileChooser fileChooser = new JFileChooser();
-        fileChooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
-        fileChooser.setFileFilter(new FileNameExtensionFilter("Json files", "json"));
-
-        int result = fileChooser.showOpenDialog(this);
-        if (result == JFileChooser.APPROVE_OPTION) {
-            File selectedFile = fileChooser.getSelectedFile();
-            if (selectedFile.getName().endsWith(".json")) {
-                openJson(selectedFile);
-            } else {
-                JOptionPane.showMessageDialog(this, "The selected file is not a json file.", "Error", JOptionPane.ERROR_MESSAGE);
-                openJSonSchedule(); // restart the file chooser
-            }
-        }
-    }
-
-    private void openCSV(File csvFile){
-        String csvData = null;
-        JFrame csvFrame = new JFrame("content of " + csvFile);
-        csvFrame.setSize(700,300);
-        csvFrame.setLocationRelativeTo(null);
-
-        try {
-            csvData = new String(Files.readAllBytes(csvFile.toPath()));
-        } catch (IOException e) {
-            LOGGER.info("Não é possivel ler o arquivo .csv selecionado!");
-        }
-
-        JPanel panel = new JPanel(new GridLayout(1, 1));
-        JTextArea  jsonText = new JTextArea(csvData);
-        panel.add(jsonText);
-
-        JScrollPane scrollPane = new JScrollPane(panel);
-        csvFrame.add(scrollPane);
-        csvFrame.add(panel);
-
-        csvFrame.setVisible(true);
-
-    }
 
 
-    private void openJson(File jsonFile)  {
-        String jsonData = null;
-        JFrame jsonInfo = new JFrame("Content of " + jsonFile.getName());
-        jsonInfo.setSize(500,500);
-        jsonInfo.setLocationRelativeTo(null);
-
-        try {
-            jsonData = new String(Files.readAllBytes(jsonFile.toPath()));
-        } catch (IOException e) {
-            LOGGER.info("Não é possivel ler o arquivo .json selecionado!");
-        }
-
-        JPanel panel = new JPanel(new GridLayout(1, 1));
-        JTextArea  jsonText = new JTextArea(jsonData);
-        panel.add(jsonText);
-
-        JScrollPane scrollPane = new JScrollPane(panel);
-        jsonInfo.add(scrollPane);
-        jsonInfo.add(panel);
-
-        jsonInfo.setVisible(true);
-
-    }
 
     private void backToMainMenu() {
         MainMenu mainMenu = new MainMenu();
